+++ index.html
@@ -1,89 +1,3 @@
-<!-- <!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="UTF-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<title>Document</title>
-		<script src="https://cdn.tailwindcss.com"></script>
-		<script>
-			tailwind.config = {
-				theme: {
-					extend: {
-						colors: {
-							clifford: "#da373d",
-						},
-					},
-				},
-			};
-		</script>
-		<script type="module" src="/JS/app.js"></script>
-		<script type="module" src="/JS/api/auth/login.js"></script>
-	</head>
-	<body>
-		<div class="flex items-center justify-center min-h-screen bg-gray-100">
-			<div
-				class="relative flex flex-col m-6 space-y-8 bg-white shadow-2xl rounded-2xl md:flex-row md:space-y-0"
-			>
-				<div class="flex flex-col justify-center p-8 md:p-14">
-					<span class="mb-3 text-4xl font-bold">Welcome back</span>
-					<span class="font-light text-gray-400 mb-8"
-						>Please enter your details</span
-					>
-					<div class="py-4">
-						<span class="mb-2 text-md">Email</span>
-						<input
-							type="text"
-							class="w-full p-2 border border-gray-300 rounded-md placeholder:font-light placeholder:text-gray-500"
-							name="email"
-							id="email"
-						/>
-					</div>
-					<div class="py-4">
-						<span class="mb-2 text-md">Password</span>
-						<input
-							type="password"
-							class="w-full p-2 border border-gray-300 rounded-md placeholder:font-light placeholder:text-gray-500"
-							name="pass"
-							id="pass"
-						/>
-					</div>
-					<div class="flex justify-between w-full py-4">
-						<div class="mr-24">
-							<input
-								type="checkbox"
-								name="checkbox"
-								id="checkbox"
-								class="mr-2"
-							/>
-							<span class="md-4">Remember for 30 days</span>
-						</div>
-						<span class="font-bold text-md">Forgot your password?</span>
-					</div>
-					<button
-						class="w-full bg-black text-white p-2 rounded-lg mb-6 hover:bg-white hover:text-black hover:border hover:border-gray-300"
-					>
-						Sign in
-					</button>
-
-					<div class="text-center text-gray-400">
-						Dont have an account?
-						<a href="/register/index.html"
-							><span class="text-bold text-black">Sign up for free</span></a
-						>
-					</div>
-				</div>
-				<div class="relative">
-					<img
-						src="/images/illustration-log-in.png"
-						alt=""
-						class="w-[400px] h-full hidden rounded-r-2xl md:block object-cover"
-					/>
-				</div>
-			</div>
-		</div>
-	</body>
-</html> -->
-
 <!DOCTYPE html>
 <html lang="en">
 	<head>
